server/requirements.txt
fastapi
uvicorn[standard]
authlib
httpx
google-auth
google-api-python-client
sqlalchemy